+++ common/src/main/java/com/railwayteam/railways/mixin_interfaces/RailwaySavedDataDuck.java
@@ -1,6 +1,6 @@
 /*
  * Steam 'n' Rails
- * Copyright (c) 2022-2024 The Railways Team
+ * Copyright (c) 2026 The Railways Team
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -16,24 +16,15 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.railwayteam.railways.mixin;
+package com.railwayteam.railways.mixin_interfaces;
 
-
-import com.simibubi.create.content.trains.entity.Carriage;
-import com.simibubi.create.content.trains.entity.Carriage.DimensionalCarriageEntity;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.ResourceKey;
-import net.minecraft.world.level.Level;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.gen.Accessor;
+import com.simibubi.create.content.trains.entity.Train;
+import net.minecraft.resources.ResourceLocation;
 
 import java.util.Map;
+import java.util.UUID;
 
-@Mixin(value = Carriage.class, remap = false)
-public interface AccessorCarriage {
-	@Accessor
-	Map<Integer, CompoundTag> getSerialisedPassengers();
-
-	@Accessor("entities")
-	Map<ResourceKey<Level>, DimensionalCarriageEntity> railways$getEntities();
+public interface RailwaySavedDataDuck {
+    Map<UUID, Train> railway$getShadowTrains();
+    Map<ResourceLocation, UUID> railways$getShadowKeys();
 }